rework build to generate bootable payload and include rkbin as submodule

+++ src/types.h
@@ -3,26 +3,26 @@
 
 #define NULL 0
 
-typedef long i64;
-typedef int i32;
+typedef long  i64;
+typedef int   i32;
 typedef short i16;
-typedef char i8;
+typedef char  i8;
 
-typedef long int64_t;
-typedef int int32_t;
+typedef long  int64_t;
+typedef int   int32_t;
 typedef short int16_t;
-typedef char int8_t;
+typedef char  int8_t;
 
-typedef unsigned long u64;
-typedef unsigned int u32;
+typedef unsigned long  u64;
+typedef unsigned int   u32;
 typedef unsigned short u16;
-typedef unsigned char u8;
+typedef unsigned char  u8;
 
-typedef unsigned long uint64_t;
-typedef unsigned int uint32_t;
+typedef unsigned long  uint64_t;
+typedef unsigned int   uint32_t;
 typedef unsigned short uint16_t;
-typedef unsigned char uint8_t;
+typedef unsigned char  uint8_t;
 
-typedef unsigned int size_t;
+typedef uint64_t size_t;
 
 #endif // TYPES_H
